+++ src/libstddjb/bitarray_testandpoke.c
@@ -1,8 +1,9 @@
 /* ISC license. */
 
+#include <sys/types.h>
 #include <skalibs/bitarray.h>
 
-int bitarray_testandpoke (register unsigned char *s, register unsigned int n, register int h)
+int bitarray_testandpoke (register unsigned char *s, register size_t n, register int h)
 {
   register unsigned char mask = 1 << (n & 7) ;
   register unsigned char c = s[n>>3] ;
