+++ src/libstddjb/bufalloc_init.c
@@ -1,9 +1,10 @@
 /* ISC license. */
 
+#include <sys/types.h>
 #include <skalibs/stralloc.h>
 #include <skalibs/bufalloc.h>
 
-void bufalloc_init (bufalloc *ba, int (*op)(int, char const *, unsigned int), int fd)
+void bufalloc_init (bufalloc *ba, int (*op)(int, char const *, size_t), int fd)
 {
   ba->x = stralloc_zero ;
   ba->op = op ;
