+++ src/libstddjb/genwrite_put_buffer.c
@@ -1,9 +1,10 @@
 /* ISC license. */
 
+#include <sys/types.h>
 #include <skalibs/buffer.h>
 #include <skalibs/genwrite.h>
 
-int genwrite_put_buffer (void *target, char const *s, unsigned int len)
+ssize_t genwrite_put_buffer (void *target, char const *s, size_t len)
 {
   register buffer *b = target ;
   return buffer_put(b, s, len) ;
