+++ src/libstddjb/genwrite_put_stralloc.c
@@ -1,10 +1,11 @@
 /* ISC license. */
 
+#include <sys/types.h>
 #include <skalibs/stralloc.h>
 #include <skalibs/genwrite.h>
 
-int genwrite_put_stralloc (void *target, char const *s, unsigned int len)
+ssize_t genwrite_put_stralloc (void *target, char const *s, size_t len)
 {
   register stralloc *sa = target ;
-  return stralloc_catb(sa, s, len) ? (int)len : -1 ;
+  return stralloc_catb(sa, s, len) ? (ssize_t)len : -1 ;
 }
