+++ src/libstddjb/unsanitize_read.c
@@ -1,10 +1,11 @@
 /* ISC license. */
 
+#include <sys/types.h>
 #include <errno.h>
 #include <skalibs/allreadwrite.h>
 #include <skalibs/error.h>
 
-int unsanitize_read (int r)
+ssize_t unsanitize_read (ssize_t r)
 {
   return r == -1 ? errno == EPIPE ? (errno = 0, 0) : -1 :
          !r ? (errno = EWOULDBLOCK, -1) : r ;
